--- a/parse.go
+++ b/parse.go
@@ -5,6 +5,7 @@ import (
 	"go/types"
 	"os"
 	"path/filepath"
+	"sync"
 
 	"golang.org/x/tools/go/packages"
 )
@@ -30,8 +31,13 @@ func ParseStruct(path string, name string, mctx *MappingContext) (types.Object, 
 	return obj, nil
 }
 
+var pcache = sync.Map{}
+
 // ParseFile parses a given Go source code file.
 func ParseFile(pkgPath string, mctx *MappingContext) (*types.Package, error) {
+	if pkg, ok := pcache.Load(pkgPath); ok {
+		return pkg.(*types.Package), nil
+	}
 	oldCwd, _ := os.Getwd()
 	pkgPath, _ = filepath.Abs(pkgPath)
 	rootPath, err := findRootPath(pkgPath)
@@ -66,6 +72,7 @@ func ParseFile(pkgPath string, mctx *MappingContext) (*types.Package, error) {
 	for _, imp := range pkg.Imports() {
 		mctx.AddImport(imp.Path())
 	}
+	pcache.Store(pkgPath, pkg)
 	return pkg, nil
 }