This PR adds an upstream change from yuin/sesame that includes a package cache during the file parsing step.

https://github.com/yuin/sesame/commit/d6811c41c413a28a56978fa8df4a970246d46c18#diff-1596bd8ceb74246828aacab827b39a33075c86baa627fbbeb7491bd31eef1169R34

This singular change greatly improves the generation time in current usages

Before:
```
> $ hyperfine -r 3 -w 1 ...        
Benchmark 1: ...
  Time (mean ± σ):     44.900 s ±  2.520 s    [User: 88.128 s, System: 30.770 s]
  Range (min … max):   43.020 s … 47.763 s    3 runs
```

After
```
> $ hyperfine -r 5 -w 2 ...
Benchmark 1: ...
  Time (mean ± σ):      5.958 s ±  0.603 s    [User: 10.956 s, System: 4.146 s]
  Range (min … max):    5.363 s …  6.770 s    5 runs
```